// backend/src/db/schema.ts
// table schema generated by introspection
import { pgTable, unique, integer, varchar } from "drizzle-orm/pg-core"
import { sql } from "drizzle-orm"

export const charts = pgTable("chart_data", {
    id: integer().primaryKey(),
    month: varchar({length: 255}).notNull(),
    windows: integer().notNull(),
    mac: integer().notNull(),
    linux: integer().notNull()
});

export const cards = pgTable("dashboard_cards", {
    id: integer().primaryKey().generatedAlwaysAsIdentity(),
    title: varchar({length: 255}).notNull(),
    value: varchar({length: 255}).notNull(),
    change: varchar({length: 255}).notNull(), 
    status:  varchar({length: 255}).notNull() 
});